requirements.txt
discord.py
requests
feedparser
python-dotenv
requests